==> analysis/02_analysis.Rmd <==
---
title: "Results"
author: "Andrew Heiss and Suparna Chaudhry"
date: "Last run: `r format(Sys.time(), '%F')`"
output: 
  html_document:
    code_folding: hide
    pandoc_args:
      - "--default-image-extension=png"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.retina = 3, fig.align = "center")
```

```{r load-libraries-data, warning=FALSE, message=FALSE}
library(tidyverse)
library(crackdownsphilanthropy)
library(rstan)
library(tidybayes)
library(broom)
library(ggstance)
library(ggdag)
library(ggraph)
library(patchwork)
library(pander)
library(scales)
library(janitor)
library(here)

source(here("analysis", "options.R"))

# Load data
results <- readRDS(here("data", "derived_data", "results_clean.rds"))
# qwraps2::lazyload_cache_dir("02_analysis_cache/html")
```


# Causal pathway

Our theory and hypotheses are laid out in the causal pathway below. Our outcomes (Y: % likely to donate and amount donated) "listen to" or respond to C (crackdown), which is our main treatment or exposure. Funding (F) and issue (I) serve as heuristics for donation and influence a government's decision to crack down on NGOs, hence the dual arrows to C and Y. 

```{r causal-dag, fig.width=4.75, fig.height=2, out.width="70%"}
theory_dag <- dagify(Y ~ I + C + F,
                     C ~ I + F,
                     outcome = "Y",
                     exposure = "C",
                     labels = c("Y" = "Outcome (Likelihood of donation & amount donated)", 
                                "C" = "Anti-NGO legal crackdown", 
                                "I" = "NGO issue area", 
                                "F" = "NGO funding source")) %>% 
  tidy_dagitty(layout = "sugiyama")

plot_dag <- ggplot(theory_dag, aes(x = x, y = y, xend = xend, yend = yend)) +
  geom_dag_point(size = 6) +
  geom_dag_edges(start_cap = circle(4, "mm"),
                 end_cap = circle(4, "mm")) +
  geom_dag_text(size = pts(6), family = "Roboto Condensed", fontface = "bold") +
  geom_dag_label_repel(aes(label = label), size = pts(6), box.padding = 0.1, 
                       direction = "x", seed = 12345,
                       family = "Roboto Condensed", fontface = "plain") +
  labs(x = NULL) +
  theme_ngos() +
  theme(panel.grid = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank())

plot_dag %T>% 
  print() %T>%
  ggsave(., filename = here("analysis", "output", "figures", "causal-path.pdf"),
         width = 4.75, height = 2, units = "in", device = cairo_pdf) %>% 
  ggsave(., filename = here("analysis", "output", "figures", "causal-path.png"),
         width = 4.75, height = 2, units = "in", type = "cairo", dpi = 300)
```


# Overview of data

## Balance of experimental conditions

```{r check-conditions, results="asis"}
results %>% count(crackdown, issue, funding) %>% 
  rename(Crackdown = crackdown, Issue = issue, Funding = funding) %>% 
  janitor::adorn_totals(., where = "row") %T>% 
  pandoc.table() %>% 
  pandoc.table.return(caption = "Balance of experimental conditions {#tbl:experimental-conditions}") %>% 
  cat(file = here("analysis", "output", "tables", "tbl-experimental-conditions.md"))
```

## Descriptive statistics table

```{r tbl-descriptive-statistics, results="asis"}
vars_to_summarize <- tribble(
  ~variable_name, ~clean_name,
  "donate_likely", "Likelihood of donation",
  "donate_likely_bin", "Likelihood of donation (binary)",
  "amount_donate", "Amount hypothetically donated ($)",
  "gender", "Gender",
  "age", "Age",
  "income", "Income",
  "education", "Education",
  "religiosity", "Frequency of attending religious services",
  "ideology", "Political views",
  "political_knowledge", "Frequency of following public affairs",
  "give_charity", "Frequency of charitable donations",
  "volunteer", "Volunteered in past 12 months",
  "favor_humanitarian", "Prior favorability towards humanitarian NGOs",
  "favor_humanitarian_bin", "Prior favorability towards humanitarian NGOs (binary)",
  "favor_human_rights", "Prior favorability towards human rights NGOs",
  "favor_human_rights_bin", "Prior favorability towards human rights NGOs (binary)",
  "favor_development", "Prior favorability towards development NGOs",
  "favor_development_bin", "Prior favorability towards development NGOs (binary)",
  "check2", "Attention check 2"
)

results_summary_stats <- results %>% 
  select(one_of(vars_to_summarize$variable_name)) %>% 
  summarise_all(~list(.)) %>% 
  pivot_longer(everything(), names_to = "variable_name", values_to = "value") %>% 
  mutate(N = value %>% map_int(~ length(na.omit(.))),
         summary = map2(.x = value, .y = variable_name, ~ md_summary_row(.x, .y))) %>% 
  left_join(vars_to_summarize, by = "variable_name") %>% 
  mutate(variable_name = factor(variable_name, 
                                levels = vars_to_summarize$variable_name, 
                                ordered = TRUE)) %>% 
  arrange(variable_name) %>% 
  select(-value, -variable_name) %>% 
  unnest(summary) %>% 
  select(Variable = clean_name, N, ` ` = spark, Details = summary)

results_summary_stats %>% 
  select(-N) %>% 
  pandoc.table.return(caption = "Descriptive statistics {#tbl:descriptive-stats}",
                      split.cell = 80, split.table = Inf) %T>% 
  cat(file = here("analysis", "output", "tables", "tbl-descriptive-stats.md")) %>%
  cat()
```

## Average likelihood and amount donated across conditions

```{r tbl-results-conditions, warning=FALSE, results="asis"}
conditions_summary <- bind_rows(group_by(results, crackdown, issue, funding) %>% nest(),
                                group_by(results, crackdown, issue) %>% nest(),
                                group_by(results, crackdown) %>% nest(),
                                results %>% nest(data = everything())) %>% 
  arrange(crackdown, issue, funding) %>% 
  mutate(summary = data %>%
           map(~ summarize(., pct_likely = table(donate_likely_bin)[["Likely"]] /
                             length(donate_likely_bin),
                           mean_donation = mean(amount_donate, na.rm = TRUE),
                           sd_donation = sd(amount_donate, na.rm = TRUE),
                           N = nrow(.)))) %>% 
  unnest(summary) %>% select(-data) %>% ungroup()

conditions_summary_clean <- conditions_summary %>% 
  mutate(funding = ifelse(is.na(funding) & !is.na(issue) , "*Total*", as.character(funding)),
         issue = ifelse(is.na(issue) & !is.na(crackdown), "*Total*", as.character(issue)),
         crackdown = ifelse(is.na(crackdown), "*Total*", as.character(crackdown))) %>% 
  group_by(crackdown) %>% 
  mutate(issue = replace(issue, duplicated(issue), NA)) %>% 
  ungroup() %>% 
  mutate(crackdown = replace(crackdown, duplicated(crackdown), NA)) %>% 
  mutate(pct_likely = percent(pct_likely)) %>% 
  rename(`Crackdown condition` = crackdown, `Issue condition` = issue,
         `Funding condition` = funding, `% likely to donate` = pct_likely,
         `Amount donated (mean)` = mean_donation, `Amount donated (sd)` = sd_donation)

conditions_summary_clean %T>% 
  pandoc.table() %>% 
  pandoc.table.return(caption = "Average likelihood and amount donated across experimental conditions {#tbl:avg-results}") %>% 
  cat(file = here("analysis", "output", "tables", "tbl-avg-results.md"))
```

\

# Visualize important variables

## Likelihood of donation

```{r fig-likelihood-bars, fig.width=9, fig.height=4, out.width="80%"}
donate_summary <- results %>% 
  count(donate_likely) %>% 
  mutate(perc = n / sum(n)) %>% 
  mutate(highlight = ifelse(donate_likely %in% c("Extremely likely", "Somewhat likely"), TRUE, FALSE))

plot_donate_summary <- ggplot(donate_summary, aes(x = n, y = donate_likely, 
                                                  fill = highlight)) +
  geom_barh(stat = "identity") +
  scale_x_continuous(sec.axis = sec_axis(~ . / sum(donate_summary$n),
                                         labels = percent_format(accuracy = 1))) +
  scale_fill_manual(values = ngo_cols("green", "blue", name = FALSE), guide = FALSE) +
  labs(x = NULL, y = NULL) +
  theme_ngos() +
  theme(panel.grid.major.y = element_blank())

plot_donate_summary %T>% 
  print() %T>%
  ggsave(., filename = here("analysis", "output", "figures", "donate_summary.pdf"),
         width = 9, height = 4, units = "in", device = cairo_pdf) %>% 
  ggsave(., filename = here("analysis", "output", "figures", "donate_summary.png"),
         width = 9, height = 4, units = "in", type = "cairo", dpi = 300)
```

## Amount donated

```{r fig-amount-bars, fig.width=9, fig.height=2.75, out.width="80%"}
plot_amount_summary <- ggplot(results, aes(x = amount_donate)) +
  geom_histogram(bins = 20, fill = ngo_cols("blue")) +
  scale_x_continuous(labels = dollar) +
  labs(x = NULL, y = NULL) +
  theme_ngos() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

plot_amount_summary %T>% 
  print() %T>%
  ggsave(., filename = here("analysis", "output", "figures", "amount_summary.pdf"),
         width = 9, height = 2.75, units = "in", device = cairo_pdf) %>% 
  ggsave(., filename = here("analysis", "output", "figures", "amount_summary.png"),
         width = 9, height = 2.75, units = "in", type = "cairo", dpi = 300)
```

\

```{r all-datasets-nested}
all_models <- tribble(
  ~title, ~df,
  "Crackdown", results,
  # Ha, cheat here by renaming the issue and funding columns to crackdown. 
  # All other tests are based on the crackdown column except these two, so
  # rather than build all sorts of convoluted arguments and functions, we just
  # rename these as issue and funding
  "Issue", mutate(results, crackdown = issue),
  "Funding", mutate(results, crackdown = funding),
  # Create all the nested conditions
  "Human rights | Crackdown", filter(results, issue == "Human rights"),
  "Humanitarian assistance | Crackdown", filter(results, issue != "Human rights"),
  "Government | Crackdown", filter(results, funding == "Government"),
  "Private | Crackdown", filter(results, funding == "Private"),
  "Human rights | Government | Crackdown", filter(results, issue == "Human rights", funding == "Government"),
  "Human rights | Private | Crackdown", filter(results, issue == "Human rights", funding == "Private"),
  "Humanitarian assistance | Government | Crackdown", filter(results, issue != "Human rights", funding == "Government"),
  "Humanitarian assistance | Private | Crackdown", filter(results, issue != "Human rights", funding == "Private")
) 
```

# Treatment effects: Likelihood of donation

This time around, we're not using interactionful regression models to calculate all these differences in groups. Instead we use real live Stan code to estimate the differences in group means and proportions! 

## Priors and models

We estimate the proportion of people responding that they'd be likely to donate to the organization with a binomial distribution, with a prior $\theta$ distribution of $\text{Beta}(5, 5)$. We build the following model in Stan:

$$
\begin{aligned}
n_{\text{group 1, group 2}} &\sim \text{Binomial}(n_{\text{total in group}}, \theta_{\text{group}}) &\text{[likelihood]}\\
\text{Difference} &= n_{\text{group 2}} - n_{\text{group 1}} &\text{[difference in proportions]} \\
n &: \text{Number likely to donate} \\
\\
\theta_{\text{group 1, group 2}} &\sim \text{Beta}(5, 5) &\text{[prior prob. of being likely to donate]}
\end{aligned}
$$

```{r priors-likely, fig.width=2.75, fig.height=1.5, out.width="50%"}
likely_theta <- ggplot(data = tibble(x = c(0, 1)), aes(x = x)) +
  geom_area(stat = "function", fun = dbeta, args = list(shape1 = 5, shape2 = 5), 
            fill = "grey80", color = "black") +
  labs(x = expression(Probability ~ of ~ being ~ likely ~ to ~ donate ~ (theta)), y = "Density") +
  annotate(geom = "label", x = 0.5, y = 1, label = "Beta(5, 5)", size = pts(9)) +
  theme_ngos(base_size = 9, density = TRUE)
likely_theta

likely_theta %T>% 
  ggsave(., filename = here("analysis", "output", "figures", "prior-likely.pdf"),
         width = 2.75, height = 1.5, units = "in", device = cairo_pdf) %>%
  ggsave(., filename = here("analysis", "output", "figures", "prior-likely.png"),
         width = 2.75, height = 1.5, units = "in", type = "cairo", dpi = 300)
```

## Differences

```{r run-likely-models, results="hide", cache=TRUE}
# This takes forever because of MCMC sampling, so this chunk is cached

# Sample from the compiled model
all_models_likely_run <- all_models %>% 
  # Simplify data for sending to Stan
  mutate(stan_data_counts = df %>% map(~ {
    df_counts <- .x %>% 
      count(crackdown, donate_likely_bin) %>% 
      group_by(crackdown) %>% 
      mutate(total = sum(n)) %>% 
      filter(donate_likely_bin == "Likely")
    
    return(list(
      n_total_1 = df_counts$total[1],
      n_total_2 = df_counts$total[2],
      n_likely_1 = df_counts$n[1],
      n_likely_2 = df_counts$n[2]
    ))
  })) %>% 
  # Run the actual model on the simplified data
  mutate(model = stan_data_counts %>% map(~ {
    donate_likely(n_total_1 = .x$n_total_1, n_total_2 = .x$n_total_2,
                  n_likely_1 = .x$n_likely_1, n_likely_2 = .x$n_likely_2,
                  control = list(max_treedepth = 15),
                  chains = CHAINS, iter = ITER, warmup = WARMUP, seed = BAYES_SEED)
  })) 

# Extract stuff
all_models_likely <- all_models_likely_run %>%
  # Extract posterior chains
  mutate(posterior_chains_long = map(model, ~ {
      .x %>% gather_draws(theta_1, theta_2, theta_diff, pct_change)
  })) %>% 
  mutate(posterior_chains_wide = map(model, ~ {
      .x %>% spread_draws(theta_1, theta_2, theta_diff, pct_change)
  })) %>% 
  # Get HDI medians in tidy form
  mutate(tidy = posterior_chains_long %>% map(~ {
    .x %>% median_hdci() %>% to_broom_names()
  }))
```

```{r plot-likely-diffs, warning=FALSE, fig.width=4.75, fig.height=5.5, out.width="90%"}
tidied_diffs_likely <- all_models_likely %>% 
  unnest(posterior_chains_long) %>% 
  filter(.variable == "theta_diff") %>% 
  mutate(category = case_when(
    str_count(title, "\\|") == 0 ~ "Level 1",
    str_count(title, "\\|") == 1 ~ "Level 2",
    str_count(title, "\\|") == 2 ~ "Level 3"
  )) %>% 
  mutate(title = str_remove(title, " \\| Crackdown"))

level1_likely <- tidied_diffs_likely %>% 
  filter(category == "Level 1") %>%
  mutate(title = recode(title, 
                        Crackdown = "Crackdown −\nNo crackdown",
                        Issue = "Humanitarian\nassistance −\nHuman rights",
                        Funding = "Private −\nGovernment\nfunding")) %>% 
  mutate(title = fct_inorder(title))

plot_diff_likely_a <- ggplot(level1_likely, aes(x = .value, y = fct_rev(title), fill = title)) +
  stat_halfeye(.width = c(0.8, 0.95)) +
  geom_vline(xintercept = 0) +
  scale_x_continuous(labels = percent_format(accuracy = 1)) +
  scale_fill_manual(values = ngo_cols(c("blue", "red", "orange"), name = FALSE), guide = FALSE) +
  labs(x = "Difference in donation likelihood", y = NULL, tag = "A") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

level2_likely <- tidied_diffs_likely %>% 
  filter(category == "Level 2") %>%
  mutate(condition = case_when(
    title %in% c("Human rights", "Humanitarian assistance") ~ "Issue",
    title %in% c("Government", "Private") ~ "Funding"
  )) %>% 
  mutate(facet_title = case_when(
    title %in% c("Human rights", "Humanitarian assistance") ~ paste(title, "issues"),
    title %in% c("Government", "Private") ~ paste(title, "funding")
  )) %>% 
  mutate(title = case_when(
    title %in% c("Humanitarian assistance", "Private") ~ "",
    TRUE ~ "Crackdown −\nNo crackdown")
  )

plot_diff_likely_b <- ggplot(filter(level2_likely, condition == "Issue"), 
                             aes(x = .value, y = fct_rev(title))) +
  stat_halfeye(.width = c(0.8, 0.95), fill = ngo_cols("red", name = FALSE)) +
  geom_vline(xintercept = 0) +
  scale_x_continuous(labels = percent_format(accuracy = 1)) +
  labs(x = NULL, y = NULL, tag = "B") +
  facet_wrap(~ facet_title, scales = "free_y") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

plot_diff_likely_c <- ggplot(filter(level2_likely, condition == "Funding"), 
                             aes(x = .value, y = fct_rev(title))) +
  stat_halfeye(.width = c(0.8, 0.95), fill = ngo_cols("orange", name = FALSE)) +
  geom_vline(xintercept = 0) +
  scale_x_continuous(labels = percent_format(accuracy = 1)) +
  labs(x = "Difference in donation likelihood", y = NULL, tag = "C") +
  facet_wrap(~ facet_title, scales = "free_y") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

level3_likely <- tidied_diffs_likely %>% 
  filter(category == "Level 3") %>%
  separate(title, c("issue", "funding"), sep = " \\| ") %>%
  mutate(issue = paste(issue, "issues"),
         funding = paste(funding, "funding")) %>% 
  mutate(facet_title = paste0(issue, "\n", funding)) %>% 
  mutate(title = case_when(
    funding == "Private funding" ~ "",
    TRUE ~ "Crackdown −\nNo crackdown")
  )

plot_diff_likely_d <- ggplot(level3_likely, aes(x = .value, y = fct_rev(title))) +
  stat_halfeye(.width = c(0.8, 0.95), fill = ngo_cols("green", name = FALSE)) +
  geom_vline(xintercept = 0) +
  scale_x_continuous(labels = percent_format(accuracy = 1)) +
  labs(x = "Difference in donation likelihood", y = NULL, tag = "D",
       caption = "Point shows posterior median; thick black lines show 80% credible interval;\nthin black lines show 95% credible interval") +
  facet_wrap(~ facet_title, scales = "free_y") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

plot_diffs_likely_all <- plot_diff_likely_a / 
  (plot_diff_likely_b / plot_diff_likely_c) / 
  plot_diff_likely_d

plot_diffs_likely_all

plot_diffs_likely_all %T>% 
  ggsave(., filename = here("analysis", "output", "figures", "likely-diffs.pdf"),
         width = 4.75, height = 5.5, units = "in", device = cairo_pdf) %>%
  ggsave(., filename = here("analysis", "output", "figures", "likely-diffs.png"),
         width = 4.75, height = 5.5, units = "in", type = "cairo", dpi = 300)
```

```{r tbl-likely-diffs-level1, warning=FALSE, message=FALSE}
tbl_likely_tidy <- all_models_likely %>% 
  unnest(tidy) %>% 
  filter(term %in% c("theta_1", "theta_2", "theta_diff", "pct_change")) %>% 
  select(title, term, estimate) %>% 
  pivot_wider(names_from = term, values_from = estimate) %>%
  mutate(category = case_when(
    str_count(title, "\\|") == 0 ~ "Level 1",
    str_count(title, "\\|") == 1 ~ "Level 2",
    str_count(title, "\\|") == 2 ~ "Level 3"
  )) 

tbl_likely_probs <- all_models_likely %>% 
  unnest(posterior_chains_long) %>% 
  filter(.variable == "theta_diff") %>% 
  group_by(title) %>% 
  summarize(p.greater0 = mean(.value > 0),
            p.less0 = mean(.value < 0),
            p.diff.not0 = ifelse(median(.value) > 0, p.greater0, p.less0)) %>% 
  ungroup()

# Save combined table for later use in manuscript
tbl_likely_tidy %>% 
  left_join(tbl_likely_probs, by = "title") %>% 
  saveRDS(here("data", "derived_data", "results_models_likely.rds"))
  
tbl_likely_1 <- tbl_likely_tidy %>% 
  left_join(tbl_likely_probs, by = "title") %>% 
  filter(category == "Level 1") %>%
  mutate(title = recode(title, 
                        Crackdown = "Crackdown − No crackdown",
                        Issue = "*Humanitarian assistance − Human rights*",
                        Funding = "*Private − Government funding*")) %>% 
  mutate_at(vars(theta_1, theta_2, theta_diff, pct_change), 
            list(~percent_format(accuracy = 0.1)(.))) %>% 
  mutate(p.diff.not0 = as.character(round(p.diff.not0, 2))) %>%
  select(Frame = title, `% likely~Treatment~` = theta_2, `% likely~Control~` = theta_1,
         `$\\Delta$` = theta_diff, `$\\%\\Delta$` = pct_change, 
         `$p(\\Delta \\neq 0)$` = p.diff.not0)
```

```{r tbl-likely-diffs-level2, warning=FALSE, message=FALSE}
tbl_likely_2 <- tbl_likely_tidy %>% 
  left_join(tbl_likely_probs, by = "title") %>% 
  filter(category == "Level 2") %>%
  mutate(title_clean = str_remove(title, " \\| Crackdown")) %>% 
  mutate(title_clean = case_when(
    title_clean %in% c("Human rights", "Humanitarian assistance") ~ paste(title_clean, "issues"),
    title_clean %in% c("Government", "Private") ~ paste(title_clean, "funding")
  )) %>% 
  mutate(title_clean = factor(title_clean, 
                              levels = c("Human rights issues", 
                                         "Humanitarian assistance issues", 
                                         "Government funding", "Private funding"))) %>% 
  arrange(title_clean) %>% 
  mutate_at(vars(theta_1, theta_2, theta_diff, pct_change), 
            list(~percent_format(accuracy = 0.1)(.))) %>% 
  mutate(p.diff.not0 = as.character(round(p.diff.not0, 2))) %>% 
  mutate(title_clean = as.character(title_clean)) %>% 
  select(`H~2a~ and H~3a~` = title_clean, 
         `% likely~Crackdown~` = theta_2, `% likely~No\\ crackdown~` = theta_1,
         `$\\Delta$` = theta_diff, `$\\%\\Delta$` = pct_change, 
         `$p(\\Delta \\neq 0)$` = p.diff.not0)
```

```{r tbl-likely-diffs-level3, warning=FALSE, message=FALSE}
tbl_likely_3 <- tbl_likely_tidy %>% 
  left_join(tbl_likely_probs, by = "title") %>% 
  filter(category == "Level 3") %>%
  mutate(title_clean = str_remove(title, " \\| Crackdown")) %>% 
  separate(title_clean, c("Issue", "Funding"), sep = " \\| ") %>% 
  mutate(Issue = factor(Issue, levels = c("Human rights", "Humanitarian assistance")),
         Funding = factor(Funding, levels = c("Government", "Private"))) %>% 
  arrange(Issue, Funding) %>% 
  mutate_at(vars(theta_1, theta_2, theta_diff, pct_change), 
            list(~percent_format(accuracy = 0.1)(.))) %>% 
  mutate(p.diff.not0 = as.character(round(p.diff.not0, 2))) %>%
  mutate(Frame = paste0(Issue, " issues, ", Funding, " funding")) %>% 
  select(`H~2a~ and H~3a~ (nested)` = Frame,
         `% likely~Crackdown~` = theta_2, `% likely~No\\ crackdown~` = theta_1,
         `$\\Delta$` = theta_diff, `$\\%\\Delta$` = pct_change, 
         `$p(\\Delta \\neq 0)$` = p.diff.not0)
```

```{r combine-likely-tables, results="asis"}
tbl_likely_2_header <- enframe(colnames(tbl_likely_2)) %>% 
  mutate(value = (value)) %>% 
  pivot_wider(names_from = name, values_from = value) %>% 
  set_names(colnames(tbl_likely_1))

tbl_likely_3_header <- enframe(colnames(tbl_likely_3)) %>% 
  mutate(value = (value)) %>% 
  pivot_wider(names_from = name, values_from = value) %>% 
  set_names(colnames(tbl_likely_1))

bind_rows(tbl_likely_1, 
          tbl_likely_2_header, 
          set_names(tbl_likely_2, colnames(tbl_likely_1)),
          tbl_likely_3_header,
          set_names(tbl_likely_3, colnames(tbl_likely_1))) %>% 
  rename(`H~1a~` = Frame) %>% 
  pandoc.table.return(caption = 'Likelihood of donation and differences in proportions in "crackdown" (treatment) and "no crackdown" (control) conditions; values represent posterior medians {#tbl:likely-diffs}',
                      justify = "lccccc") %T>%
  cat() %>%
  cat(file = here("analysis", "output", "tables", "tbl-likely-diffs.md"))
```

\

# Treatment effects: Amount donated

## Priors and models

Following [John Kruschke's "Bayesian Estimation Supersedes the t-test (BEST)" procedure](http://www.indiana.edu/~kruschke/BEST/BEST.pdf), we estimate means for each group with a t-distribution. We use the following priors for the distribution parameters:

$$
\begin{aligned}
x_{\text{group 1, group 2}} &\sim \text{Student } t(\nu, \mu, \sigma) &\text{[likelihood]}\\
\text{Difference} &= x_{\text{group 2}} - x_{\text{group 1}} &\text{[difference in means]} \\
x &: \text{Mean amount donated} \\
\\
\nu &\sim \text{Exponential}(1 / 29) &\text{[prior normality]} \\
\mu_{\text{group 1, group 2}} &\sim \mathcal{N}(\bar{x}_{\text{group 1, group 2}}, 10) &\text{[prior donation mean per group]}\\
\sigma_{\text{group 1, group 2}} &\sim \text{Cauchy}(0, 1)&\text{[prior donation sd per group]}
\end{aligned}
$$

```{r priors-amounts, fig.width=8, fig.height=1.5, out.width="100%"}
amount_nu <- ggplot(data = tibble(x = c(0, 200)), aes(x = x)) +
  geom_area(stat = "function", fun = dexp, args = list(rate = 1/29), 
            fill = "grey80", color = "black") +
  labs(x = expression(Normality ~ parameter ~ (nu)), y = "Density") +
  annotate(geom = "label", x = 100, y = 0.009, 
           label = "Exponential(1/29)", size = pts(9)) +
  theme_ngos(base_size = 9, density = TRUE)

amount_mu <- ggplot(data = tibble(x = c(0, 100)), aes(x = x)) +
  geom_area(stat = "function", fun = dnorm, args = list(mean = 50, sd = 10),
            fill = "grey80", color = "black") +
  scale_x_continuous(breaks = c(seq(0, 100, 25)),
                     labels = c("−$50", "−$25", "Group average", "+$25", "+$50")) +
  annotate(geom = "label", x = 50, y = 0.01, label = "N(bar(x), 10)", 
           parse = TRUE, size = pts(9)) +
  labs(x = expression(Average ~ donated ~ (mu)), y = NULL) +
  theme_ngos(base_size = 9, density = TRUE)

amount_sigma <- ggplot(data = tibble(x = c(0, 10)), aes(x = x)) +
  geom_area(stat = "function", fun = dcauchy, args = list(location = 0, scale = 1),
            fill = "grey80", color = "black") +
  scale_x_continuous(labels = dollar) +
  annotate(geom = "label", x = 5, y = 0.08, label = "Cauchy(0, 1)", size = pts(9)) +
  labs(x = expression(SD ~ donated ~ (sigma)), y = NULL) +
  theme_ngos(base_size = 9, density = TRUE)

amount_priors <- amount_nu + amount_mu + amount_sigma
amount_priors

amount_priors %T>% 
  ggsave(., filename = here("analysis", "output", "figures", "prior-amount.pdf"),
         width = 8, height = 1.5, units = "in", device = cairo_pdf) %>%
  ggsave(., filename = here("analysis", "output", "figures", "prior-amount.png"),
         width = 8, height = 1.5, units = "in", type = "cairo", dpi = 300)
```


## Differences

```{r run-amount-models, results="hide", cache=TRUE, warning=FALSE}
# Sample from the compiled model
all_models_amount_run <- all_models %>% 
  # Simplify data for sending to Stan
  mutate(stan_data = df %>% map(~ {
    select(.x, amount = amount_donate, group = crackdown)
  })) %>% 
  # Run the actual model on the simplified data
  mutate(model = stan_data %>% map(~ {
    amount_donated_best(df = .x, chains = CHAINS, iter = ITER, 
                        warmup = WARMUP, seed = BAYES_SEED)
  })) 

# Extract stuff
all_models_amount <- all_models_amount_run %>% 
  # Extract posterior chains
  mutate(posterior_chains_long = map2(model, stan_data, ~ {
    .x %>% 
      recover_types(.y) %>% 
      gather_draws(mu[group], sigma[group], mu_diff, pct_change, cohen_d, cles, nu, log10nu)
  })) %>% 
  mutate(posterior_chains_wide = map2(model, stan_data, ~ {
    .x %>% 
      recover_types(.y) %>% 
      spread_draws(mu[group], sigma[group], mu_diff, pct_change, cohen_d, cles, nu, log10nu)
  })) %>% 
  # Get HDI medians in tidy form
  mutate(tidy = posterior_chains_long %>% map(~ {
    .x %>% median_hdci() %>% to_broom_names()
  }))
```

```{r plot-amount-diffs, warning=FALSE, fig.width=4.75, fig.height=5.5, out.width="90%"}
all_models_amount <- all_models_amount %>% 
  # The `group` and `tidy` columns in each of these nested models contain
  # ordered factors, but we can't combine them with unnest() directly because
  # they have different levels (crackdown, issue, and funding), so we need to
  # make them unordered factors before unnesting
  mutate(posterior_chains_long = map(
    posterior_chains_long, ~mutate(., group = factor(group, ordered = FALSE)))
  ) %>% 
  mutate(tidy = map(
    tidy, ~mutate(., group = factor(group, ordered = FALSE)))
  )

tidied_diffs_amount <- all_models_amount %>% 
  unnest(posterior_chains_long) %>% 
  filter(.variable == "mu_diff") %>% 
  mutate(category = case_when(
    str_count(title, "\\|") == 0 ~ "Level 1",
    str_count(title, "\\|") == 1 ~ "Level 2",
    str_count(title, "\\|") == 2 ~ "Level 3"
  )) %>% 
  mutate(title = str_remove(title, " \\| Crackdown"))

level1_amount <- tidied_diffs_amount %>% 
  filter(category == "Level 1") %>%
  mutate(title = recode(title, 
                        Crackdown = "Crackdown −\nNo crackdown",
                        Issue = "Humanitarian\nassistance −\nHuman rights",
                        Funding = "Private −\nGovernment\nfunding")) %>% 
  mutate(title = fct_inorder(title))

plot_diff_amount_a <- ggplot(level1_amount, aes(x = .value, y = fct_rev(title), fill = title)) +
  geom_vline(xintercept = 0) +
  stat_halfeye(.width = c(0.8, 0.95)) +
  scale_x_continuous(labels = dollar) +
  scale_fill_manual(values = ngo_cols(c("blue", "red", "orange"), name = FALSE), guide = FALSE) +
  labs(x = "Difference in amount donated", y = NULL, tag = "A") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

level2_amount <- tidied_diffs_amount %>% 
  filter(category == "Level 2") %>%
  mutate(condition = case_when(
    title %in% c("Human rights", "Humanitarian assistance") ~ "Issue",
    title %in% c("Government", "Private") ~ "Funding"
  )) %>% 
  mutate(facet_title = case_when(
    title %in% c("Human rights", "Humanitarian assistance") ~ paste(title, "issues"),
    title %in% c("Government", "Private") ~ paste(title, "funding")
  )) %>% 
  mutate(title = case_when(
    title %in% c("Humanitarian assistance", "Private") ~ "",
    TRUE ~ "Crackdown −\nNo crackdown")
  )

plot_diff_amount_b <- ggplot(filter(level2_amount, condition == "Issue"), 
                             aes(x = .value, y = fct_rev(title))) +
  stat_halfeye(.width = c(0.8, 0.95), fill = ngo_cols("red", name = FALSE)) +
  geom_vline(xintercept = 0) +
  scale_x_continuous(labels = dollar) +
  labs(x = NULL, y = NULL, tag = "B") +
  facet_wrap(~ facet_title, scales = "free_y") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

plot_diff_amount_c <- ggplot(filter(level2_amount, condition == "Funding"), 
                             aes(x = .value, y = fct_rev(title))) +
  stat_halfeye(.width = c(0.8, 0.95), fill = ngo_cols("orange", name = FALSE)) +
  geom_vline(xintercept = 0) +
  scale_x_continuous(labels = dollar) +
  labs(x = "Difference in amount donated", y = NULL, tag = "C") +
  facet_wrap(~ facet_title, scales = "free_y") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

level3_amount <- tidied_diffs_amount %>% 
  filter(category == "Level 3") %>%
  separate(title, c("issue", "funding"), sep = " \\| ") %>%
  mutate(issue = paste(issue, "issues"),
         funding = paste(funding, "funding")) %>% 
  mutate(facet_title = paste0(issue, "\n", funding)) %>% 
  mutate(title = case_when(
    funding == "Private funding" ~ "",
    TRUE ~ "Crackdown −\nNo crackdown")
  )

plot_diff_amount_d <- ggplot(level3_amount, aes(x = .value, y = fct_rev(title))) +
  stat_halfeye(.width = c(0.8, 0.95), fill = ngo_cols("green", name = FALSE)) +
  geom_vline(xintercept = 0) +
  scale_x_continuous(labels = dollar) +
  labs(x = "Difference in amount donated", y = NULL, tag = "D",
       caption = "Point shows posterior median; thick black lines show 80% credible interval;\nthin black lines show 95% credible interval") +
  facet_wrap(~ facet_title, scales = "free_y") +
  theme_ngos(base_size = 8) +
  theme(panel.grid.major.y = element_blank())

plot_diffs_amounts_all <- plot_diff_amount_a / 
  (plot_diff_amount_b / plot_diff_amount_c) / 
  plot_diff_amount_d

plot_diffs_amounts_all

plot_diffs_amounts_all %T>% 
  ggsave(., filename = here("analysis", "output", "figures", "amount-diffs.pdf"),
         width = 4.75, height = 5.5, units = "in", device = cairo_pdf) %>%
  ggsave(., filename = here("analysis", "output", "figures", "amount-diffs.png"),
         width = 4.75, height = 5.5, units = "in", type = "cairo", dpi = 300)
```

```{r tbl-amount-diffs-level1, warning=FALSE, message=FALSE}
tbl_amounts_tidy <- all_models_amount %>% 
  unnest(tidy) %>% 
  filter(term %in% c("mu", "mu_diff", "pct_change")) %>% 
  group_by(title) %>% 
  mutate(group_id = 1:n()) %>% 
  mutate(term = case_when(
    term == "mu" ~ paste0(term, "_", group_id),
    TRUE ~ term
  )) %>% 
  select(title, term, estimate) %>% 
  pivot_wider(names_from = term, values_from = estimate) %>% 
  mutate(category = case_when(
    str_count(title, "\\|") == 0 ~ "Level 1",
    str_count(title, "\\|") == 1 ~ "Level 2",
    str_count(title, "\\|") == 2 ~ "Level 3"
  )) %>% 
  ungroup()

tbl_amounts_probs <- all_models_amount %>% 
  unnest(posterior_chains_long) %>% 
  filter(.variable == "mu_diff") %>% 
  group_by(title) %>% 
  summarize(p.greater0 = mean(.value > 0),
            p.less0 = mean(.value < 0),
            p.diff.not0 = ifelse(median(.value) > 0, p.greater0, p.less0)) %>% 
  ungroup()

# Save combined table for later use in manuscript
tbl_amounts_tidy %>% 
  left_join(tbl_amounts_probs, by = "title") %>% 
  saveRDS(here("data", "derived_data", "results_models_amount.rds"))

tbl_amount_1 <- tbl_amounts_tidy %>% 
  left_join(tbl_amounts_probs, by = "title") %>% 
  filter(category == "Level 1") %>%
  mutate(title = recode(title, 
                        Crackdown = "Crackdown − No crackdown",
                        Issue = "*Humanitarian assistance − Human rights*",
                        Funding = "*Private − Government funding*")) %>% 
  mutate_at(vars(mu_1, mu_2, mu_diff, p.diff.not0), list(~as.character(round(., 2)))) %>% 
  mutate(pct_change = percent_format(accuracy = 0.1)(pct_change)) %>%
  select(`Frame` = title, `Amount~Treatment~` = mu_2, `Amount~Control~` = mu_1,
         `$\\Delta$` = mu_diff, `$\\%\\Delta$` = pct_change, 
         `$p(\\Delta \\neq 0)$` = p.diff.not0)
```

```{r tbl-amount-diffs-level2, warning=FALSE, message=FALSE}
tbl_amount_2 <- tbl_amounts_tidy %>% 
  left_join(tbl_amounts_probs, by = "title") %>% 
  filter(category == "Level 2") %>%
  mutate(title_clean = str_remove(title, " \\| Crackdown")) %>% 
  mutate(title_clean = case_when(
    title_clean %in% c("Human rights", "Humanitarian assistance") ~ paste(title_clean, "issues"),
    title_clean %in% c("Government", "Private") ~ paste(title_clean, "funding")
  )) %>% 
  mutate(title_clean = factor(title_clean, 
                              levels = c("Human rights issues", 
                                         "Humanitarian assistance issues", 
                                         "Government funding", "Private funding"))) %>% 
  arrange(title_clean) %>% 
  mutate_at(vars(mu_1, mu_2, mu_diff, p.diff.not0), list(~as.character(round(., 2)))) %>% 
  mutate(pct_change = percent_format(accuracy = 0.1)(pct_change)) %>%
  mutate(title_clean = as.character(title_clean)) %>% 
  select(`H~2b~ and H~3b~` = title_clean, `Amount~Crackdown~` = mu_2, `Amount~No\\ crackdown~` = mu_1,
         `$\\Delta$` = mu_diff, `$\\%\\Delta$` = pct_change, 
         `$p(\\Delta \\neq 0)$` = p.diff.not0)
```

```{r tbl-amount-diffs-level3, warning=FALSE, message=FALSE}
tbl_amount_3 <- tbl_amounts_tidy %>% 
  left_join(tbl_amounts_probs, by = "title") %>% 
  filter(category == "Level 3") %>%
  mutate(title_clean = str_remove(title, " \\| Crackdown")) %>% 
  separate(title_clean, c("Issue", "Funding"), sep = " \\| ") %>% 
  mutate(Issue = factor(Issue, levels = c("Human rights", "Humanitarian assistance")),
         Funding = factor(Funding, levels = c("Government", "Private"))) %>% 
  arrange(Issue, Funding) %>% 
  mutate(Frame = paste0(Issue, " issues, ", Funding, " funding")) %>% 
  mutate_at(vars(mu_1, mu_2, mu_diff, p.diff.not0), list(~as.character(round(., 2)))) %>% 
  mutate(pct_change = percent_format(accuracy = 0.1)(pct_change)) %>%
  select(`H~2b~ and H~3b~ (nested)` = Frame,
         `Amount~Crackdown~` = mu_2, `Amount~No\\ crackdown~` = mu_1,
         `$\\Delta$` = mu_diff, `$\\%\\Delta$` = pct_change, 
         `$p(\\Delta \\neq 0)$` = p.diff.not0)
```

```{r combine-amount-tables, results="asis"}
tbl_amount_2_header <- enframe(colnames(tbl_amount_2)) %>% 
  mutate(value = (value)) %>% 
  pivot_wider(names_from = name, values_from = value) %>% 
  set_names(colnames(tbl_amount_1))

tbl_amount_3_header <- enframe(colnames(tbl_amount_3)) %>% 
  mutate(value = (value)) %>% 
  pivot_wider(names_from = name, values_from = value) %>% 
  set_names(colnames(tbl_amount_1))

bind_rows(tbl_amount_1, 
          tbl_amount_2_header, 
          set_names(tbl_amount_2, colnames(tbl_amount_1)),
          tbl_amount_3_header,
          set_names(tbl_amount_3, colnames(tbl_amount_1))) %>% 
  rename(`H~1b~` = Frame) %>% 
  pandoc.table.return(caption = 'Mean values and differences in means for amount donated in "crackdown" (treatment) and "no crackdown" (control) conditions; values represent posterior medians {#tbl:amount-diffs}',
                      justify = "lccccc") %T>%
  cat() %>%
  cat(file = here("analysis", "output", "tables", "tbl-amount-diffs.md"))
```

## Effect size

```{r amount-effect-size, warning=FALSE, out.width="80%"}
# Effect size
# (μ₁ - μ₂) / sqrt( (σ₁² + σ₂²) / 2)
tidied_eff_size_amount <- all_models_amount %>% 
  unnest(posterior_chains_long) %>% 
  filter(.variable == "cohen_d") %>% 
  mutate(category = case_when(
    str_count(title, "\\|") == 0 ~ "Level 1",
    str_count(title, "\\|") == 1 ~ "Level 2",
    str_count(title, "\\|") == 2 ~ "Level 3"
  )) %>% 
  mutate(title = str_remove(title, " \\| Crackdown")) %>% 
  mutate(title = fct_inorder(title))

effect_sizes <- tribble(
  ~size, ~x_end,
  "Small", 0.2,
  "Medium", 0.5,
  "Large", 0.8
) %>% 
  mutate(x_start = -x_end)

ggplot(tidied_eff_size_amount, aes(x = .value, y = fct_rev(title))) +
  stat_halfeye(.width = c(0.8, 0.95)) +
  geom_vline(xintercept = 0, size = 1) +
  geom_vline(xintercept = c(0.2, -0.2), linetype = "dotted", size = 0.5) +
  geom_vline(xintercept = c(0.5, -0.5), linetype = "dotted", size = 0.5) +
  geom_vline(xintercept = c(0.5, -0.5), linetype = "dotted", size = 0.5) +
  coord_cartesian(xlim = c(-1, 1)) +
  labs(x = "Cohen's d (effect size)", y = NULL) +
  theme_ngos()
```

\

# Original computing environment

<button data-toggle="collapse" data-target="#sessioninfo" class="btn btn-primary btn-md btn-info">Here's what we used the last time we built this page</button>

<div id="sessioninfo" class="collapse">

```{r show-session-info, echo=TRUE, width=100}
writeLines(readLines(file.path(Sys.getenv("HOME"), ".R/Makevars")))

devtools::session_info()
```

</div>
